+++ pymetrica/codebase_parser/codebase_parser.py
@@ -5,7 +5,7 @@ from pathlib import Path
 
 from pymetrica.models import Code, Codebase
 from pymetrica.utils import is_comment_line, is_logical_line_of_code, log
-from pymetrica.utils.settings import Configuration
+from pymetrica.utils.settings import Config
 
 # TODO ignore folders inside .gitignore
 
@@ -21,7 +21,7 @@ def parse_codebase(dir_path: str) -> Codebase:  # pylint: disable=too-many-local
 
     for path in base.rglob("*.py"):
         if is_excluded(path, base):
-            log.warning(
+            log.info(
                 f"parse_codebase.{path = } excluded by pyproject.toml configuration",
             )
             continue
@@ -30,8 +30,8 @@ def parse_codebase(dir_path: str) -> Codebase:  # pylint: disable=too-many-local
         try:
             tree = ast.parse(source)
         except SyntaxError as e:
-            log.warning(f"parse_codebase.SyntaxError.{path = }")
-            log.warning(f"parse_codebase.SyntaxError: {e = }")
+            log.debug(f"parse_codebase.SyntaxError.{path = }")
+            log.debug(f"parse_codebase.SyntaxError: {e = }")
             continue
         total_classes_definitions += sum(
             isinstance(n, ast.ClassDef) for n in ast.walk(tree)
@@ -89,13 +89,13 @@ def parse_codebase(dir_path: str) -> Codebase:  # pylint: disable=too-many-local
 
 def is_excluded(path: Path, root: Path) -> bool:
     rel_path = path.relative_to(root).as_posix()
-    return any(fnmatch(rel_path, p) for p in Configuration.exclude)
+    return any(fnmatch(rel_path, p) for p in Config.exclude)
 
 
 def get_base_path(dir_path: str) -> Path:
     base = Path(dir_path).absolute()
     if dir_path == ".":
-        log.warning(
+        log.info(
             f"parse_codebase.{dir_path = }, which resolves to {base}",
         )
         src_folder = Path(str(base) + os.sep + "src")
@@ -112,7 +112,7 @@ def get_base_path(dir_path: str) -> Path:
 def check_possible_bases(base: Path, folders: list[Path]) -> Path:
     for folder in folders:
         if folder.exists() and folder.is_dir():
-            log.warning(
+            log.info(
                 f"parse_codebase: {folder} exists and is a directory, "
                 f"using it as base path instead of {base}",
             )