+++ pymetrica/metric_calculators/abstract_lines_of_code/first_pass.py
@@ -55,7 +55,8 @@ def gather_loc_and_classes(files: list[Code]) -> PreliminaryResults:  # pylint: 
             # if isinstance(node, ast.Call):
             #     file_abs.add(node.lineno)
 
-        total_aloc += len(file_abs)
+        file.aloc_number = len(file_abs)
+        total_aloc += file.aloc_number
 
         # Measure classes
         for node in tree.body:
@@ -77,7 +78,7 @@ def gather_loc_and_classes(files: list[Code]) -> PreliminaryResults:  # pylint: 
                         try:
                             bases.append(ast.unparse(b))
                         except:  # noqa: E722  # pylint: disable=bare-except
-                            log.warning(
+                            log.debug(
                                 f"gather_loc_and_classes.unparse_failed.{ast.dump(b) = }",
                             )
                 # Treat ABC, ABCMeta, Protocol, and Enum as abstract