+++ pymetrica/metric_calculators/instability/instability_metric.py
@@ -1,22 +1,37 @@
 import json
 
 from pymetrica.models import Metric, Results
+from pymetrica.utils import log
 
 
 class InstabilityResults(Results):
     instability: dict[str, float]
 
-    def get_dict(self) -> dict[str, float]:
+    @property
+    def dict_(self) -> dict[str, float]:
         return self.instability
 
-    def get_json(self) -> str:
-        return json.dumps(self.get_dict())
+    @property
+    def json_(self) -> str:
+        return json.dumps(self.dict_)
 
-    def get_summary(self) -> str:
+    @property
+    def summary(self) -> str:
         summary = "\nInstability for:\n"
         for layer, instability in self.instability.items():
             summary += f"  {layer}: {instability}\n"
         return summary
 
+    @property
+    def fail_message(self) -> str:
+        log.debug("Instability metric does not have a fail threshold.")
+        return ""
 
-class InstabilityMetric(Metric[InstabilityResults]): ...
+    @property
+    def exceeds_threshold(self) -> bool:
+        log.debug("Instability metric does not have a fail threshold.")
+        return False
+
+
+class InstabilityMetric(Metric[InstabilityResults]):
+    exit_code: int = 0