+++ pymetrica/metric_calculators/primitive_obsession/get_po.py
@@ -6,22 +6,22 @@ from pymetrica.codebase_parser import parse_codebase
 from pymetrica.report_generators import REPORTS_MAPPING
 from pymetrica.utils import run_profiler
 
-from .mc_calculator import MaintainabilityCostCalculator
+from .po_calculator import PrimitiveObsessionCalculator
 
-mc_calculator: MaintainabilityCostCalculator = MaintainabilityCostCalculator()
+po_calculator: PrimitiveObsessionCalculator = PrimitiveObsessionCalculator()
 
 
 @click.command()
 @click.argument("dir_path")
 @click.option("-rt", "--report-type", type=str, default="BASIC_TERMINAL")
 @run_profiler
-def mc(
+def po(
     dir_path: str,
     report_type: str,
 ) -> int:
     codebase = parse_codebase(dir_path)
-    mc_metric = mc_calculator.calculate_metric(codebase)
-    report_generator = REPORTS_MAPPING[report_type]([mc_metric])
+    po_metric = po_calculator.calculate_metric(codebase)
+    report_generator = REPORTS_MAPPING[report_type]([po_metric])
     report = report_generator.long_report
 
     click.echo(report.content)