+++ pymetrica/run_all.py
@@ -3,29 +3,23 @@ import sys
 import click
 
 from pymetrica.codebase_parser import parse_codebase
-from pymetrica.metric_calculators.abstract_lines_of_code.aloc_calculator import (
-    AlocCalculator,
-)
-from pymetrica.metric_calculators.cyclomatic_complexity.cc_calculator import (
-    CCCalculator,
-)
-from pymetrica.metric_calculators.halstead_volume.hv_calculator import (
-    HalsteadVolumeCalculator,
-)
-from pymetrica.metric_calculators.instability.instability_calculator import (
-    InstabilityCalculator,
-)
-from pymetrica.metric_calculators.maintainability_cost.mc_calculator import (
-    MaintainabilityCostCalculator,
-)
 from pymetrica.models.metric import Metric, Results
 from pymetrica.report_generators.reports_mapping import REPORTS_MAPPING
-from pymetrica.utils import Configuration
+
+from .metric_calculators import (
+    AlocCalculator,
+    CCCalculator,
+    HalsteadVolumeCalculator,
+    InstabilityCalculator,
+    MaintainabilityCostCalculator,
+    PrimitiveObsessionCalculator,
+)
 
 aloc_calculator: AlocCalculator = AlocCalculator()
 cc_calculator: CCCalculator = CCCalculator()
 hv_calculator: HalsteadVolumeCalculator = HalsteadVolumeCalculator()
 mc_calculator: MaintainabilityCostCalculator = MaintainabilityCostCalculator()
+po_calculator: PrimitiveObsessionCalculator = PrimitiveObsessionCalculator()
 instability_calculator: InstabilityCalculator = InstabilityCalculator()
 
 
@@ -47,54 +41,16 @@ def run_all(
     metrics.append(aloc_calculator.calculate_metric(codebase))
     metrics.append(cc_calculator.calculate_metric(codebase))
     metrics.append(hv_calculator.calculate_metric(codebase))
+    metrics.append(po_calculator.calculate_metric(codebase))
     metrics.append(mc_calculator.calculate_metric(codebase))
     metrics.append(instability_calculator.calculate_metric(codebase))
 
-    report_generator = REPORTS_MAPPING[report_type]()
-    if long_report:
-        click.echo(report_generator.generate_report(metrics))
-        return
-    click.echo(report_generator.generate_short_report(metrics))
+    report_generator = REPORTS_MAPPING[report_type](metrics)
 
-    exit_status = 0
-    if (
-        Configuration.aloc_fail_threshold != 0
-        and metrics[0].results.aloc_percentage > Configuration.aloc_fail_threshold  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-    ):
-        click.echo(
-            f"ALOC percentage {metrics[0].results.aloc_percentage:.2f}% exceeds "  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-            f"the fail threshold of {Configuration.aloc_fail_threshold}%",
-            err=True,
-        )
-        exit_status += 1
-    if (
-        Configuration.cc_fail_threshold != 0
-        and metrics[1].results.lloc_per_cc > Configuration.cc_fail_threshold  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-    ):
-        click.echo(
-            f"LLOC per CC {metrics[1].results.lloc_per_cc:.2f} exceeds the "  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-            f"fail threshold of {Configuration.cc_fail_threshold}",
-            err=True,
-        )
-        exit_status += 10
-    if (
-        Configuration.hv_fail_threshold != 0
-        and metrics[2].results.hv_per_lloc > Configuration.hv_fail_threshold  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-    ):
-        click.echo(
-            f"Halstead Volume per LLOC {metrics[2].results.hv_per_lloc:.2f} "  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-            f"exceeds the fail threshold of {Configuration.hv_fail_threshold}",
-            err=True,
-        )
-        exit_status += 100
-    if (
-        Configuration.mc_fail_threshold != 0
-        and metrics[3].results.maintainability_cost > Configuration.mc_fail_threshold  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-    ):
-        click.echo(
-            f"Maintainability Cost {metrics[3].results.maintainability_cost:.2f} "  # type: ignore[attr-defined]  # pylint: disable=line-too-long
-            f"exceeds the fail threshold of {Configuration.mc_fail_threshold}",
-            err=True,
-        )
-        exit_status += 1000
-    sys.exit(exit_status)
+    if long_report:
+        report = report_generator.long_report
+    else:
+        report = report_generator.short_report
+
+    click.echo(report.content)
+    sys.exit(report.exit_status)