+++ pymetrica/utils/__init__.py
@@ -2,10 +2,10 @@ from .is_comment import is_comment_line
 from .is_lloc import is_logical_line_of_code
 from .logging import log
 from .profiler import run_profiler
-from .settings import Configuration
+from .settings import Config
 
 __all__ = [
-    "Configuration",
+    "Config",
     "is_comment_line",
     "is_logical_line_of_code",
     "log",